backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
google-auth==2.23.4
google-auth-oauthlib==1.2.3
google-api-python-client==2.108.0
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7

